+++ DataSet.h
@@ -4,4 +4,4 @@
 
 using DataSet_t = std::vector<float>;
 
-std::vector<DataSet_t> makeRandomDataSets(size_t dataSets, size_t features, const DecisionTree_t& tree);
+std::vector<DataSet_t> makeRandomDataSets(uint32_t dataSets, uint32_t features, const DecisionTree_t& tree);
